+++ src/models/users/services/singInUser.go
@@ -15,20 +15,6 @@ import (
 	"github.com/rodrigoRVSN/beeus-api/src/models/users/entities"
 )
 
-func CreateUser(payload *entities.User) error {
-	errors := fieldsValidator.ValidateStruct(payload)
-
-	if errors != nil {
-		return errors
-	}
-
-	payload.Password = hash.HashPassword(payload.Password)
-
-	database.DB.Db.Create(&payload)
-
-	return nil
-}
-
 func SignInUser(payload *userDtos.SignInInput) (*userDtos.SignInOutput, error) {
 	errors := fieldsValidator.ValidateStruct(payload)
